requirements.txt
Django>=5.0,<6.0
djangorestframework
gunicorn
python-decouple
whitenoise
drf-spectacular
djangorestframework-simplejwt
django-filter
celery
redis
django-celery-beat
django-celery-results
